add if contitions for specific bloks, fix sqlplus query

+++ oracle/install_docker_oracle.sh
@@ -6,13 +6,25 @@ echo "-----------------------------------------------------------------"
 
 . "../lib/install_methods.sh"
 USER_HOME=$(get_home_dir)
-ORACLE_CONTAINER_NAME="oracle19c-sql"
+ORACLE_CONTAINER_NAME="oracle19c"
 ORACLE_IMAGE_NAME="oracle/database:19.3.0-ee"
 ONLY_SQL=true
-PATH_TO_ORACLE_ZIP="${1:-$HOME}/LINUX.X64_193000_db_home.zip"
+PATH_TO_ORACLE_ZIP="$HOME/LINUX.X64_193000_db_home.zip"
+SKIP_ZIP_PULL=false;
+
+# -------- Parse arguments --------
+while getopts "sp:" opt; do
+  case $opt in
+    p) PATH_TO_ORACLE_ZIP="$OPTARG" ;;
+    s) SKIP_ZIP_PULL=true ;;
+    *) echo "Usage: $0 [-p oracle_zip_path]" && exit 1 ;;
+  esac
+done
 
 if [[ "$ONLY_SQL" == false ]]; then
-    echo "Checking for Docker installation..."
+    echo "---------------------------------------------------"
+    echo "      Checking for Docker installation..."
+    echo "---------------------------------------------------"
     if ! command -v docker &> /dev/null; then
         read -p "Install Docker? (y/N): " -n 1 -r
         echo
@@ -36,16 +48,36 @@ if [[ "$ONLY_SQL" == false ]]; then
     else
         echo "Docker is already installed."
     fi
-    echo "Docker installation check complete."
 
-    if [[ -d "$USER_HOME/docker-images" ]]; then
-        echo "Removing existing docker-images directory..."
-        rm -rf "$USER_HOME/docker-images"
+    echo  -e "---------------------------------------------------"
+    echo "      Docker installation check complete."
+    echo "---------------------------------------------------"
+
+    if ! $SKIP_ZIP_PULL; then
+        echo -e "---------------------------------------------------"
+        echo "          Pull Oracle Docker repositpory"
+        echo "---------------------------------------------------"
+        if [[ -d "$USER_HOME/docker-images" ]]; then
+            echo "Removing existing docker-images directory..."
+            rm -rf "$USER_HOME/docker-images"
+        fi
+        echo "Pulling Oracle Docker scripts..."
+        git clone https://github.com/oracle/docker-images.git "$USER_HOME/docker-images"
+
+        if [[ -f $PATH_TO_ORACLE_ZIP ]]; then
+            cp "$PATH_TO_ORACLE_ZIP" "$USER_HOME/docker-images/OracleDatabase/SingleInstance/dockerfiles/19.3.0/"
+        else
+            echo "$PATH_TO_ORACLE_ZIP does't exist"
+            return 1
+        fi
+        echo -e "---------------------------------------------------"
+        echo "     Pull Oracle Docker repositpory complete"
+        echo "---------------------------------------------------"
     fi
-    echo "Pulling Oracle Docker scripts..."
-    git clone https://github.com/oracle/docker-images.git "$USER_HOME/docker-images"
-    cp "$PATH_TO_ORACLE_ZIP" "$USER_HOME/docker-images/OracleDatabase/SingleInstance/dockerfiles/19.3.0/"
-    echo "Oracle Docker scripts pulled successfully."
+
+    echo -e "---------------------------------------------------"
+    echo "              Docker inage creation"
+    echo "---------------------------------------------------"
     if ! image_exists "$ORACLE_IMAGE_NAME"; then
         echo "Building Oracle Database Docker image..."
         sudo $USER_HOME/docker-images/OracleDatabase/SingleInstance/dockerfiles/buildContainerImage.sh -v 19.3.0 -e #todo support for other versions
@@ -53,6 +85,9 @@ if [[ "$ONLY_SQL" == false ]]; then
     else
         echo "Oracle Database Docker image already exists."
     fi
+    echo -e "---------------------------------------------------"
+    echo "         Docker inage creation complete"
+    echo "---------------------------------------------------"
 
     echo -e "Creating docker container..."
     sudo docker run -d --name $ORACLE_CONTAINER_NAME -p 1521:1521 -p 5500:5500 -v oracle:/opt/oracle:rw  -e ORACLE_PWD=manager $ORACLE_IMAGE_NAME
@@ -67,18 +102,16 @@ if is_container_running "$ORACLE_CONTAINER_NAME"; then
     DB_USER="sys"
     DB_PASSWORD="manager"
     PDB_SERVICE="ORCLCDB"
+    SQL_SCRIPT="/tmp/pdb_script.sql"
+
 
     echo "Creating pluggable database and user..."
-    sudo docker cp pdb_script.sql $ORACLE_CONTAINER_NAME:/tmp/pdb_script.sql
-    sudo docker exec -i "$ORACLE_CONTAINER_NAME" bash -c "
-    sqlplus -s ${DB_USER}/${DB_PASSWORD}@${PDB_SERVICE} as sysdba <<EOF
-    SET ECHO ON
-    SET FEEDBACK ON
-    SET SERVEROUTPUT ON
-    @/tmp/pdb_script.sql
-    EXIT
-    EOF
-    "
+    sudo docker cp pdb_script.sql $ORACLE_CONTAINER_NAME:$SQL_SCRIPT
+    sudo docker exec -i "$ORACLE_CONTAINER_NAME" sqlplus ${DB_USER}/${DB_PASSWORD}@${PDB_SERVICE} as sysdba @"$SQL_SCRIPT" && echo "Success" || echo "Failed"
+
+
+else
+    echo "Cannot find container $ORACLE_CONTAINER_NAME"
 fi
